Split out a base 'Wikidata' class
There's still too much in here, but we can clean that up later.
